test: add a ULA 2MHz switching page to the teletext disc
T8 is a MODE 4 page with MA13 set so the SAA5050 sees the bitmap's
bytes, and a cycle counted raster loop that flips the teletext select
bit six times a scanline, 11 cycles (5.5 cells) apart. Equal boxes mean
the ULA switches its output at 2MHz; 5 and 6 alternating means it only
switches on cell boundaries. Its reference image pins the fix for #766.

The page pokes the CRTC directly because MOS 3.20 folds the *TV
interlace flag into R8 writes made through VDU 23. Rendering a page is
now shared between capture-refs.js and the integration test, since T8
runs until a key rather than stopping at a GET.

Also documents T7, which the previous commit left out of the README.
