+++ tests/hardware/teletext/build.asm
@@ -8,3 +8,4 @@ PUTBASIC "src/t4.bas", "T4"
 PUTBASIC "src/t5.bas", "T5"
 PUTBASIC "src/t6.bas", "T6"
 PUTBASIC "src/t7.bas", "T7"
+PUTBASIC "src/t8.bas", "T8"
