+++ tests/hardware/teletext/capture-refs.js
@@ -1,4 +1,4 @@
-// Renders each test page under jsbeeb and writes t1.png to t6.png into the
+// Renders each test page under jsbeeb and writes t1.png to t8.png into the
 // refs directory beside this script, for comparison against a photograph of
 // the same page on real hardware.
 //
@@ -9,29 +9,14 @@
 
 import { mkdirSync, writeFileSync } from "fs";
 import path from "node:path";
-import { fileURLToPath } from "node:url";
-import { MachineSession } from "../../../src/machine-session.js";
-
-const ScriptDir = path.dirname(fileURLToPath(import.meta.url));
-const Disc = path.join(ScriptDir, "teletext-tests.ssd");
-const OutputDir = path.join(ScriptDir, "refs");
-const Pages = ["T1", "T2", "T3", "T4", "T5", "T6", "T7"];
+import { Pages, RefDir, renderPage } from "./render-page.js";
 
 async function main() {
-    mkdirSync(OutputDir, { recursive: true });
+    mkdirSync(RefDir, { recursive: true });
 
     for (const page of Pages) {
-        const session = new MachineSession("Master");
-        await session.initialise();
-        await session.boot();
-        session.loadDisc(Disc);
-        await session.type(`CHAIN "${page}"`);
-        // Each page ends at a GET, so the OS reaching the keyboard is the page being finished.
-        await session.runUntilPrompt();
-        await session.runFrames(1);
-        const file = path.join(OutputDir, `${page.toLowerCase()}.png`);
-        writeFileSync(file, await session.screenshotActive());
-        session.destroy();
+        const file = path.join(RefDir, `${page.toLowerCase()}.png`);
+        writeFileSync(file, await renderPage(page));
         console.log(`wrote ${file}`);
     }
 }
