+++ tests/integration/teletext-hardware.js
@@ -1,37 +1,15 @@
 import { describe, it } from "vitest";
 import assert from "assert";
 import path from "node:path";
-import { fileURLToPath } from "node:url";
 import sharp from "sharp";
 import pixelmatch from "pixelmatch";
-import { MachineSession } from "../../src/machine-session.js";
+import { Pages, RefDir, renderPage } from "../hardware/teletext/render-page.js";
 
-// The reference images were photographed against a real BBC Master 128 running
-// MOS 3.20 and agreed with it on every page, so they pin behaviour we have
-// measured rather than behaviour we have merely chosen. See
-// tests/hardware/teletext/README.md for what each page tests.
-const HardwareDir = path.join(path.dirname(fileURLToPath(import.meta.url)), "../hardware/teletext");
-const Disc = path.join(HardwareDir, "teletext-tests.ssd");
-const RefDir = path.join(HardwareDir, "refs");
+// The T1 to T6 reference images were photographed against a real BBC Master 128
+// running MOS 3.20 and agreed with it on every page, so they pin behaviour we have
+// measured rather than behaviour we have merely chosen. T7 and T8 await a
+// photograph. See tests/hardware/teletext/README.md for what each page tests.
 const OutputDir = "tests/integration/output";
-const Pages = ["T1", "T2", "T3", "T4", "T5", "T6", "T7"];
-
-async function renderPage(page) {
-    const session = new MachineSession("Master");
-    try {
-        await session.initialise();
-        await session.boot();
-        session.loadDisc(Disc);
-        await session.type(`CHAIN "${page}"`);
-        // Each page ends at a GET, so the OS reaching the keyboard is the page having
-        // finished drawing. This keeps the flash phase reproducible too.
-        await session.runUntilPrompt();
-        await session.runFrames(1);
-        return await session.screenshotActive();
-    } finally {
-        session.destroy();
-    }
-}
 
 async function compareToReference(page, actualPng) {
     const name = page.toLowerCase();
